# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AHS Portal API - Allied Health College, Nishtar Medical University
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PROGRAMS
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List all programs
      responses:
        "200":
          description: Programs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"

    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get program by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"

    patch:
      operationId: updateProgram
      tags: [programs]
      summary: Update program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"

    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ADMISSION SESSIONS
...
  /quotas:
    get:
      operationId: listQuotas
      tags: [quotas]
      summary: List quota categories
      responses:
        "200":
          description: Quotas list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotaCategory"

    post:
      operationId: createQuota
      tags: [quotas]
      summary: Create quota category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotaBody"
      responses:
        "201":
          description: Created quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaCategory"
  /quotas/{id}:
    patch:
      operationId: updateQuota
      tags: [quotas]
      summary: Update quota
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotaBody"
      responses:
        "200":
          description: Updated quota
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotaCategory"

    delete:
      operationId: deleteQuota
      tags: [quotas]
      summary: Delete quota
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # APPLICATIONS
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"

  # PAYMENT CHALLANS
...
  /verifications:
    get:
      operationId: listVerifications
      tags: [verification]
      summary: List verification decisions
      parameters:
        - name: applicationId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Verifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationDecision"

    post:
      operationId: createVerification
      tags: [verification]
      summary: Submit verification decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVerificationBody"
      responses:
        "201":
          description: Verification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDecision"

  # NOTICES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, email, fullName, role, isActive, createdAt, updatedAt]
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
      required: [email, password, fullName, role]
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [users, total, page, limit]

    # STUDENT PROFILE
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fatherName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        cnic:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        domicile:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        matricMarks:
          type: ["integer", "null"]
        matricTotal:
          type: ["integer", "null"]
        interMarks:
          type: ["integer", "null"]
        interTotal:
          type: ["integer", "null"]
        interYear:
          type: ["integer", "null"]
        interBoard:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, createdAt, updatedAt]
    UpdateProfileBody:
      type: object
      properties:
        fatherName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        cnic:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        domicile:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        matricMarks:
          type: ["integer", "null"]
        matricTotal:
          type: ["integer", "null"]
        interMarks:
          type: ["integer", "null"]
        interTotal:
          type: ["integer", "null"]
        interYear:
          type: ["integer", "null"]
        interBoard:
          type: ["string", "null"]

    # PROGRAMS
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        duration:
          type: string
        seats:
          type: integer
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, code, duration, seats, isActive, createdAt]
    CreateProgramBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        duration:
          type: string
        seats:
          type: integer
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name, code, duration, seats, isActive]

    # ADMISSION SESSIONS
...
    QuotaCategory:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        percentage:
          type: number
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, code, percentage, isActive, createdAt]
    CreateQuotaBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        percentage:
          type: number
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name, code, percentage, isActive]

    # APPLICATIONS
...
    VerificationDecision:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        officerId:
          type: integer
        status:
          type: string
        remarks:
          type: ["string", "null"]
        verifiedAt:
          type: string
        createdAt:
          type: string
      required: [id, applicationId, officerId, status, verifiedAt, createdAt]
    CreateVerificationBody:
      type: object
      properties:
        applicationId:
          type: integer
        status:
          type: string
        remarks:
          type: ["string", "null"]
      required: [applicationId, status]

    # NOTICES
...